Task #5077: Revert the two merges from trunk

+++ branches/Cobalt-Task5077/RTCP/Cobalt/GPUProc/src/cuda/Kernels/BeamFormerTransposeKernel.h
@@ -42,7 +42,7 @@ namespace LOFAR
       struct Parameters : Kernel::Parameters
       {
         Parameters(const Parset& ps);
-        unsigned nrTABs;
+        size_t nrTABs;
       };
 
       enum BufferType
